feat: FruitStore 싱글톤패턴 적용, Fruit열거형, Juice열거형 파일분리
