+++ JuiceMaker/JuiceMaker/Controller/ViewController.swift
@@ -10,12 +10,10 @@ class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
         
         let myJuiceMaker = JuiceMaker()
         myJuiceMaker.makeJuice(.strawberryBananaJuice, total: 1)
         myJuiceMaker.makeJuice(.strawberryBananaJuice, total: 1)
-        
         
     }
