+++ JuiceMaker/JuiceMaker/Model/Fruit.swift
@@ -1,0 +1,7 @@
+enum Fruit {
+    case strawberry
+    case banana
+    case pineapple
+    case kiwi
+    case mango
+}
