+++ JuiceMaker/JuiceMaker/Model/JuiceMakerError.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 enum JuiceMakerError: Error {
     case stockShortage, noSuchFruit
 }
